Ledger Monolith deploys to VM and successfully runs Spring Boot Java App
