+++ src/ledgermonolith/init/db/tables.sql
@@ -12,22 +12,22 @@
 -- See the License for the specific language governing permissions and
 -- limitations under the License.
 
-CREATE TABLE TRANSACTIONS (
-    TRANSACTION_ID BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    FROM_ACCT CHAR(10) NOT NULL,
-    TO_ACCT CHAR(10) NOT NULL,
-    FROM_ROUTE CHAR(9) NOT NULL,
-    TO_ROUTE CHAR(9) NOT NULL,
-    AMOUNT INT NOT NULL,
-    TIMESTAMP TIMESTAMP NOT NULL
+CREATE TABLE transactions (
+    transaction_id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+    from_acct CHAR(10) NOT NULL,
+    to_acct CHAR(10) NOT NULL,
+    from_route CHAR(9) NOT NULL,
+    to_route CHAR(9) NOT NULL,
+    amount INT NOT NULL,
+    timestamp TIMESTAMP NOT NULL
 );
 -- index account number/routing number pairs
-CREATE INDEX ON TRANSACTIONS (FROM_ACCT, FROM_ROUTE, TIMESTAMP);
-CREATE INDEX ON TRANSACTIONS (TO_ACCT, TO_ROUTE, TIMESTAMP);
+CREATE INDEX ON transactions (from_acct, from_route, timestamp);
+CREATE INDEX ON transactions (to_acct, to_route, timestamp);
 -- append only ledger; prevent updates or deletes
 CREATE RULE PREVENT_UPDATE AS
-  ON UPDATE TO TRANSACTIONS
+  ON UPDATE TO transactions
   DO INSTEAD NOTHING;
 CREATE RULE PREVENT_DELETE AS
-  ON DELETE TO TRANSACTIONS
+  ON DELETE TO transactions
   DO INSTEAD NOTHING;
