+++ src/ledgermonolith/src/main/java/anthos/samples/financedemo/ledgermonolith/LedgerMonolithController.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2020, Google LLC.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package anthos.samples.financedemo.ledgermonolith;
+
+import java.util.logging.Logger;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * REST service to retrieve the current balance for the authenticated user.
+ */
+@RestController
+public final class LedgerMonolithController {
+
+    private static final Logger LOGGER =
+            Logger.getLogger(LedgerMonolithController.class.getName());
+
+    public static final String READINESS_CODE = "ok";
+
+    @Value("${VERSION}")
+    private String version;
+
+   /**
+     * Version endpoint.
+     *
+     * @return  service version string
+     */
+    @GetMapping("/version")
+    public ResponseEntity<String> version() {
+        return new ResponseEntity<String>(version, HttpStatus.OK);
+    }
+
+    /**
+     * Readiness probe endpoint.
+     *
+     * @return HTTP Status 200 if server is ready to receive requests.
+     */
+    @GetMapping("/ready")
+    @ResponseStatus(HttpStatus.OK)
+    public ResponseEntity<String> readiness() {
+        return new ResponseEntity<String>(READINESS_CODE, HttpStatus.OK);
+    }
+}
